services/order-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python